# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MOSAP3 Procurement API",
        "description": "Documentação oficial da API do Sistema de Procurement MOSAP3.\n * \n *      Roles:\n *      - **Admin**: Acesso total ao sistema.\n *      - **Procurement Technician**: Gestão de cotações e fornecedores.\n *      - **Public (Fornecedor)**: Acesso via token único enviado por email.",
        "contact": {
            "email": "admin@mosap3.ao"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/api/quotation/{token}": {
            "get": {
                "tags": [
                    "Link Público (Email)"
                ],
                "summary": "Visualizar Pedido",
                "description": "Permite ao fornecedor ver os detalhes do pedido usando o token do email.",
                "operationId": "1a0be3ebc240642b2acd82cb6028cb10",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pedido recuperados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "quotation_supplier": {
                                            "type": "object"
                                        },
                                        "existing_response": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quotation/{token}/submit": {
            "post": {
                "tags": [
                    "Link Público (Email)"
                ],
                "summary": "Submeter Proposta",
                "operationId": "9d8886b8ca6faaa8060e828035a1339c",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "delivery_date",
                                    "delivery_days",
                                    "payment_terms"
                                ],
                                "properties": {
                                    "delivery_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-02-15"
                                    },
                                    "delivery_days": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "payment_terms": {
                                        "type": "string",
                                        "example": "50% na encomenda, 50% na entrega"
                                    },
                                    "observations": {
                                        "type": "string",
                                        "example": "Frete incluso"
                                    },
                                    "proposal_file": {
                                        "description": "Documento da proposta (PDF, DOC, DOCX - max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proposta enviada com sucesso"
                    }
                }
            }
        },
        "/api/quotation/{token}/decline": {
            "post": {
                "tags": [
                    "Link Público (Email)"
                ],
                "summary": "Declinar Participação",
                "operationId": "434a134e7c017ff6e272ee1b561b2db0",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participação apenas declinada"
                    }
                }
            }
        },
        "/api/quotation/{token}/attachments/{index}": {
            "get": {
                "tags": [
                    "Link Público (Email)"
                ],
                "summary": "Descarregar Anexo do Pedido de Cotação",
                "description": "Permite ao fornecedor descarregar um documento anexado ao pedido de cotação. Acesso via token do email, sem autenticação necessária.",
                "operationId": "53c31a381e1d8229db1711b8178aaace",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token de acesso do fornecedor (enviado por email)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "index",
                        "in": "path",
                        "description": "Índice do anexo na lista de documentos (começa em 0)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documento retornado com sucesso",
                        "content": {
                            "application/pdf": {},
                            "application/msword": {},
                            "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {},
                            "image/jpeg": {},
                            "image/png": {}
                        }
                    },
                    "404": {
                        "description": "Anexo ou ficheiro não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}